add getTrainingSessionAPI to trainingSession.ts
